Spelling Corrections
Spelling mistakes checked and updated. Names and titles capitalized. 

For Radical Dreamers, we decided to put the title of the game; Chrono Cross, in the heading tag and the name of the song; Radical Dreamers in the tag below it as this is the format for all of the other songs. 

When I started making the changes, I rather absent mindedly changed a spelling error on line 50:

"name": "gurren_lagan_overlay",

to:

"name": "gurren_lagann_overlay",

I immediately realised my mistake and changed it back. However, I just though I'd let you know just in case.